requirements.txt
annotated-types==0.6.0
anyio==4.6.2.post1
certifi==2024.12.14
cffi==1.16.0
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cryptography==42.0.5
cycler==0.12.1
deprecation==2.1.0
exceptiongroup==1.2.2
fastapi==0.115.5
fonttools==4.55.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.6
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
kiwisolver==1.4.7
matplotlib==3.8.4
numpy==1.26.4
packaging==24.2
pandas==2.2.2
pillow==10.4.0
postgrest==0.16.4
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
realtime==1.0.3
six==1.16.0
sniffio==1.3.1
starlette==0.41.2
supabase==2.4.6
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.32.1
websockets==12.0
requests==2.32.3
